# vanlife-frontend/docs/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "OpenAPI definition",
      "version": "v0"
    },
    "servers": [
      {
        "url": "http://localhost:9000",
        "description": "Generated server url"
      }
    ],
    "paths": {
      "/api/users/{id}": {
        "get": {
          "tags": [
            "user-controller"
          ],
          "operationId": "getUserById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "user-controller"
          ],
          "operationId": "updateUser",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "user-controller"
          ],
          "operationId": "deleteUser",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "patch": {
          "tags": [
            "user-controller"
          ],
          "operationId": "patchUser",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "/api/users": {
        "get": {
          "tags": [
            "user-controller"
          ],
          "operationId": "getAllUsers",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "user-controller"
          ],
          "operationId": "createUser",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "/api/tags": {
        "post": {
          "tags": [
            "tag-controller"
          ],
          "operationId": "createTag",
          "parameters": [
            {
              "name": "name",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "predefined",
              "in": "query",
              "required": true,
              "schema": {
                "type": "boolean"
              }
            },
            {
              "name": "description",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        }
      },
      "/api/register": {
        "post": {
          "tags": [
            "registration-controller"
          ],
          "operationId": "registerUser",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/tags/{name}": {
        "get": {
          "tags": [
            "tag-controller"
          ],
          "operationId": "getTagByName",
          "parameters": [
            {
              "name": "name",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/tags/{id}": {
        "delete": {
          "tags": [
            "tag-controller"
          ],
          "operationId": "deleteTag",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/locations/{locationId}/tags/{tagId}": {
        "delete": {
          "tags": [
            "location-controller"
          ],
          "operationId": "removeTagFromLocation",
          "parameters": [
            {
              "name": "locationId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            },
            {
              "name": "tagId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/Location"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string"
            },
            "picture": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "roles": {
              "type": "string"
            },
            "active": {
              "type": "boolean"
            }
          }
        },
        "Tag": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "name": {
              "type": "string"
            },
            "predefined": {
              "type": "boolean"
            },
            "description": {
              "type": "string"
            }
          }
        },
        "RegistrationRequest": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          }
        },
        "Comment": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "text": {
              "type": "string"
            },
            "user": {
              "$ref": "#/components/schemas/User"
            },
            "location": {
              "$ref": "#/components/schemas/Location"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "Location": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "name": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "latitude": {
              "type": "number",
              "format": "double"
            },
            "longitude": {
              "type": "number",
              "format": "double"
            },
            "address": {
              "type": "string"
            },
            "mapServiceId": {
              "type": "string"
            },
            "mapUrl": {
              "type": "string"
            },
            "ratings": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Rating"
              }
            },
            "comments": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Comment"
              }
            },
            "createdBy": {
              "$ref": "#/components/schemas/User"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "updatedAt": {
              "type": "string",
              "format": "date-time"
            },
            "amenities": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "photoUrls": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "safetyRating": {
              "type": "number",
              "format": "double"
            },
            "accessibility": {
              "type": "string"
            },
            "cost": {
              "type": "string"
            },
            "popularity": {
              "type": "integer",
              "format": "int32"
            },
            "tags": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Tag"
              }
            }
          }
        },
        "Rating": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "score": {
              "type": "integer",
              "format": "int32"
            },
            "user": {
              "$ref": "#/components/schemas/User"
            },
            "location": {
              "$ref": "#/components/schemas/Location"
            }
          }
        },
...
      }
    }
  }